requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
supabase==2.3.4
cohere==4.57
pinecone-client==3.1.0
groq==0.4.2
python-dotenv==1.0.1
pydantic>=2.7.0